Refaktorointi: oppijat.ts pilkottu pieniksi
