+++ valpas-web/src/state/apitypes/henkilo.ts
@@ -1,0 +1,17 @@
+import { ISODate, Oid } from "../common"
+
+export type HenkilöLaajatTiedot = {
+  oid: Oid
+  hetu?: string
+  syntymäaika?: ISODate
+  etunimet: string
+  sukunimi: string
+  turvakielto: boolean
+}
+
+export type HenkilöSuppeatTiedot = {
+  oid: Oid
+  syntymäaika?: ISODate
+  etunimet: string
+  sukunimi: string
+}
