+++ valpas-web/src/state/apitypes/organisaatiot.ts
@@ -1,0 +1,11 @@
+import { LocalizedString, Oid } from "../common"
+
+export type Oppilaitos = {
+  oid: Oid
+  nimi: LocalizedString
+}
+
+export type Toimipiste = {
+  oid: Oid
+  nimi: LocalizedString
+}
