+++ valpas-web/src/views/oppija/OppijanOppivelvollisuustiedot.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { InfoTable, InfoTableRow } from "../../components/tables/InfoTable"
 import { t } from "../../i18n/i18n"
-import { OppijaHakutilanteillaLaajatTiedot } from "../../state/oppijat"
+import { OppijaHakutilanteillaLaajatTiedot } from "../../state/apitypes/oppija"
 import { formatNullableDate } from "../../utils/date"
 
 export type OppijanOppivelvollisuustiedotProps = {
